backend/python_server/requirements.txt
numpy==1.24.4
scikit-learn==1.3.2
scipy==1.10.1
nltk==3.8.1
spacy==3.7.2
underthesea==1.3.5
fastapi==0.110.1
uvicorn[standard]==0.27.1
pandas==2.0.3
requests==2.31.0
torch>=1.13.1
ftfy
regex
tqdm
Pillow
open-clip-torch>=2.32.0
underthesea==1.3.5